+++ merino/utils/api/metrics.py
@@ -1,0 +1,25 @@
+"""A utility module for merino API metrics."""
+
+from collections import Counter
+from aiodogstatsd import Client
+
+from merino.providers.base import BaseProvider, BaseSuggestion
+
+
+def emit_suggestions_per_metrics(
+    metrics_client: Client,
+    suggestions: list[BaseSuggestion],
+    searched_providers: list[BaseProvider],
+) -> None:
+    """Emit metrics for suggestions per request and suggestions per request by provider."""
+    metrics_client.histogram("suggestions-per.request", value=len(suggestions))
+
+    suggestion_counter = Counter(suggestion.provider for suggestion in suggestions)
+
+    for provider in searched_providers:
+        provider_name = provider.name
+        suggestion_count = suggestion_counter[provider_name]
+        metrics_client.histogram(
+            f"suggestions-per.provider.{provider_name}",
+            value=suggestion_count,
+        )
