+++ tests/unit/utils/api/test_query_params.py
@@ -11,7 +11,7 @@ from fastapi import Request
 
 from merino.middleware import ScopeKey
 from merino.middleware.geolocation import Location
-from merino.utils.api_v1 import (
+from merino.utils.api.query_params import (
     get_accepted_languages,
     refine_geolocation_for_suggestion,
     validate_suggest_custom_location_params,
